More improvements to markdown readme system
The Markdown Rendering system now supports code blocks, quote blocks, images, mote headings, different colors for headings, lines, lines under headings 1 and 2, better indentation on lists and more. All with caching of expensive parsing operations.

I also included handling for when the Readme.MD is on the main folder of the project (as is usually the case with github) and make the handling of linking to other .md use the same method as handling regular links so ideally when using links for .md on github it can work in the inspector mode as well.